refactor code - changed file name auth -> services
