+++ lib/providers/auth_providers.dart
@@ -1,7 +1,7 @@
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import '../Auth/google_sign_in.dart';
+import '../Services/google_sign_in.dart';
 
 final authenticationProvider = Provider<GoogleAuth>((ref) => GoogleAuth());
